--- a/example/icnb/gen.go
+++ b/example/icnb/gen.go
@@ -1,0 +1,3 @@
+package icnb
+
+//go:generate ../../bin/modelgen --extended -p icnb -o . ovn-ic-nb.ovsschema