--- a/example/icsb/gen.go
+++ b/example/icsb/gen.go
@@ -1,0 +1,3 @@
+package icsb
+
+//go:generate ../../bin/modelgen --extended -p icsb -o . ovn-ic-sb.ovsschema