--- a/example/nb/gen.go
+++ b/example/nb/gen.go
@@ -1,0 +1,3 @@
+package nb
+
+//go:generate ../../bin/modelgen --extended -p nb -o . ovn-nb.ovsschema