--- a/example/sb/gen.go
+++ b/example/sb/gen.go
@@ -1,0 +1,3 @@
+package sb
+
+//go:generate ../../bin/modelgen --extended -p sb -o . ovn-sb.ovsschema