--- a/example/vtep/gen.go
+++ b/example/vtep/gen.go
@@ -1,0 +1,3 @@
+package vtep
+
+//go:generate ../../bin/modelgen --extended -p vtep -o . vtep.ovsschema