Fix #365
```
$ benchstat bench.out.old bench.out.new
goos: linux
goarch: amd64
pkg: github.com/ovn-org/libovsdb/client
cpu: Intel(R) Core(TM) i7-1065G7 CPU @ 1.30GHz
                                                            │ bench.out.old │            bench.out.new             │
                                                            │    sec/op     │    sec/op     vs base                │
APIList/predicate_returns_none-4                               3.297m ±  5%   1.072m ±  7%  -67.48% (p=0.000 n=10)
APIList/predicate_returns_all-4                                74.05m ± 13%   86.59m ±  7%  +16.95% (p=0.029 n=10)
APIList/predicate_on_an_arbitrary_condition-4                  12.39m ± 26%   10.30m ± 12%  -16.85% (p=0.000 n=10)
APIList/predicate_matches_name-4                               4.111m ± 14%   1.433m ± 19%  -65.14% (p=0.000 n=10)
APIList/by_index,_no_predicate-4                               37.06µ ± 21%   37.29µ ± 14%        ~ (p=0.971 n=10)
APIListMultiple/multiple_results_one_at_a_time_with_Get-4      6.706m ± 21%   6.582m ± 26%        ~ (p=0.739 n=10)
APIListMultiple/multiple_results_in_a_batch_with_WhereAny-4    3.104m ±  3%   3.513m ± 11%  +13.19% (p=0.003 n=10)
Update1-4                                                      107.9µ ±  7%   122.1µ ± 10%  +13.14% (p=0.002 n=10)
Update2-4                                                      166.6µ ±  3%   180.9µ ± 12%   +8.60% (p=0.015 n=10)
Update3-4                                                      223.9µ ±  4%   237.3µ ± 18%   +5.98% (p=0.023 n=10)
Update5-4                                                      337.3µ ±  2%   350.4µ ±  5%   +3.89% (p=0.011 n=10)
Update8-4                                                      519.5µ ±  2%   542.6µ ± 11%   +4.45% (p=0.005 n=10)
geomean                                                        1.173m         1.014m        -13.56%

                                                            │ bench.out.old │            bench.out.new             │
                                                            │     B/op      │     B/op      vs base                │
APIList/predicate_returns_none-4                               908.6Ki ± 0%   664.4Ki ± 0%  -26.87% (p=0.000 n=10)
APIList/predicate_returns_all-4                                14.88Mi ± 0%   14.65Mi ± 0%   -1.54% (p=0.000 n=10)
APIList/predicate_on_an_arbitrary_condition-4                  2.468Mi ± 0%   2.229Mi ± 0%   -9.67% (p=0.000 n=10)
APIList/predicate_matches_name-4                               910.3Ki ± 0%   666.1Ki ± 0%  -26.82% (p=0.000 n=10)
APIList/by_index,_no_predicate-4                               82.00Ki ± 0%   82.00Ki ± 0%        ~ (p=1.000 n=10)
APIListMultiple/multiple_results_one_at_a_time_with_Get-4      1.440Mi ± 0%   1.440Mi ± 0%        ~ (p=0.810 n=10)
APIListMultiple/multiple_results_in_a_batch_with_WhereAny-4    713.0Ki ± 0%   713.0Ki ± 0%        ~ (p=0.481 n=10)
Update1-4                                                      32.00Ki ± 0%   32.01Ki ± 0%        ~ (p=0.171 n=10)
Update2-4                                                      49.60Ki ± 0%   49.60Ki ± 0%        ~ (p=0.753 n=10)
Update3-4                                                      67.27Ki ± 0%   67.28Ki ± 0%   +0.02% (p=0.009 n=10)
Update5-4                                                      102.7Ki ± 0%   102.6Ki ± 0%        ~ (p=0.183 n=10)
Update8-4                                                      155.3Ki ± 0%   155.3Ki ± 0%        ~ (p=0.403 n=10)
geomean                                                        359.6Ki        338.0Ki        -6.00%

                                                            │ bench.out.old │             bench.out.new             │
                                                            │   allocs/op   │  allocs/op   vs base                  │
APIList/predicate_returns_none-4                              20010.00 ± 0%    10.00 ± 0%  -99.95% (p=0.000 n=10)
APIList/predicate_returns_all-4                                 230.1k ± 0%   210.1k ± 0%   -8.69% (p=0.000 n=10)
APIList/predicate_on_an_arbitrary_condition-4                   43.36k ± 0%   23.36k ± 0%  -46.12% (p=0.000 n=10)
APIList/predicate_matches_name-4                              20032.00 ± 0%    32.00 ± 0%  -99.84% (p=0.000 n=10)
APIList/by_index,_no_predicate-4                                 32.00 ± 0%    32.00 ± 0%        ~ (p=1.000 n=10) ¹
APIListMultiple/multiple_results_one_at_a_time_with_Get-4       22.02k ± 0%   22.02k ± 0%        ~ (p=1.000 n=10)
APIListMultiple/multiple_results_in_a_batch_with_WhereAny-4     11.51k ± 0%   11.51k ± 0%        ~ (p=1.000 n=10)
Update1-4                                                        855.0 ± 0%    855.0 ± 0%        ~ (p=1.000 n=10) ¹
Update2-4                                                       1.353k ± 0%   1.353k ± 0%        ~ (p=1.000 n=10) ¹
Update3-4                                                       1.851k ± 0%   1.851k ± 0%        ~ (p=1.000 n=10) ¹
Update5-4                                                       2.844k ± 0%   2.844k ± 0%        ~ (p=1.000 n=10) ¹
Update8-4                                                       4.329k ± 0%   4.329k ± 0%        ~ (p=1.000 n=10) ¹
geomean                                                         5.551k        1.624k       -70.75%
¹ all samples are equal
```