@halfcrazy 2023-09-19
CPU profile from bench test
![new](https://github.com/ovn-org/libovsdb/assets/4113173/a8f41938-f969-4a8e-ba41-75864ac1031a)

![old](https://github.com/ovn-org/libovsdb/assets/4113173/e82a056e-1c7f-45ba-820a-0659a848338d)