@halfcrazy 2023-09-22
> Seems generally the right thing to do... Most (all?) uses of WhereCache() already have a model that the predicate uses.

Emm, I found it difficult to get table information from the predicate callback function since `model.Model` is just an interface, not a struct. I cannot create a `model.Model` instance and call Table().